shop menu

+++ drally.h
@@ -8,7 +8,7 @@
 #include <SDL2/SDL.h>
 
 typedef unsigned long long  qword;
-typedef unsigned int 	    dword;
+typedef unsigned long 	    dword;
 typedef unsigned short 	    word;
 typedef unsigned char 	    byte;
 typedef void *              pointer;
